hubble-relay: Add relay status message
This adds a status message to the observer service implementation
of hubble-relay. The status message is added as a new variant of
the GetFlowsResponse, therefore older clients will silently ignore
it.

The status message contains

  - The total number of known nodes in the cluster
  - The number of nodes which were queried to satisfy the request (which
    in the current implementation is equal to the above, but this will
    likely be optimized in the future).
  - The number of nodes which responded successfully during processing.
  - The names of the nodes which were unresponsive or which returned an
    error during processing.

For non-follow-mode requests, the status message it sent at the end of
the request, i.e. once all flows have been received from the peers. In
follow-mode, the status message it sent every 30 seconds.

The current implementation of GetFlows does not try to reconnect to
peers during a request, therefore the current implementation does not
have to deal with peers becoming available again.

Fixes: #11360

Signed-off-by: Sebastian Wicki <sebastian@isovalent.com>