test: Cleanup default namespace before each Context()
Many tests still deploy into the default namespace and sometimes leak
resources. This makes it difficult to reliably re-use clusters for
testing. Clean the default namespace from various deployment resources.

This is specifically not using `delete all --all` so it can be used with
the default namespace. It also doesn't touch services, serviceaccounts,
secrets, clusterroles, ... with the assumption that these can always be
safely overwritten.

Signed-off-by: Thomas Graf <thomas@cilium.io>